AddreAddressBech32 as a Branded Type instead of a Newtype
