+++ packages/runtime/core/package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "@marlowe.io/language-core-v1": "0.3.0-beta-rc1",
+    "@marlowe.io/adapter": "0.3.0-beta-rc1",
     "fp-ts": "^2.16.1",
     "io-ts": "2.2.21",
     "newtype-ts": "0.3.5"
